solution for converter with JUNIT solution

+++ app/src/main/java/com/example/salis/currencyconverter/CurrencyConverter.java
@@ -3,7 +3,7 @@ package com.example.salis.currencyconverter;
 public class CurrencyConverter {
 
 
-    private static final int CURRENCY_CONVERSION = 200;
+    private static final double CURRENCY_CONVERSION = 200;
 
     public static double convertINRToIDR(int currency) {
         return currency * CURRENCY_CONVERSION;
